add manager for debug run, and delete old script
